claim_verification/requirements.txt
langchain
langchain-openai
langchain-perplexity
openai
python-dotenv
langchain-core
perplexityai